changes made on 2/2/16
A bunch of tools added
1.isInside
2.isOn
3.isInsideAndNotOn
4.normalize
5.simplifyVector
6.makeVector

I also created a testing executable in my make file so that I might
eventually be able to test all changes I make to see if it affected all
or any of my code and how it functions. Although there are still many
tests that still need to be added, the dream is still there.

Changes to the MyFeatureDetector.cpp
I removed a couple of redundant functions like drawShape that took in
different parameters.
I changed drawCircle to be the main drawing circle function.
I removed a lot of the previous work I did on that one.
Still needs touching up though for it to work.