MAJOR UPDATE
1. This version inserts circles in a completely new way and correctly
inserts them using a different definition for what it means to be inside
a shape.

2. A popular algorithm for finding if a point is inside a shape or not
is generating a ray in all directions of a point and if the point is
inside the shape, the rays will intersect the shape an odd number of
times whereas if the point is inside the shape, the rays will intersect
the shape an even number of times. However, this way of checking to see
if a point is inside a shape or not is highly inefficient and often
produces poor results due to the limits of the types of shapes it can
check.

I propose 2 different definitions for "inside a shape".

A. The Fly Definition

A point is inside a shape if there is no possible way for the fly
(point) to get out of the shape without crossing the shape.

B. The Maze Definition

A point is inside a shape if there is no possible way for the point to
get out of the shape in a straight line from the point without crossing
the shape.

I now use either definition depending on the situation for what I needed
to define as "inside the shape".

3. The program will now run by itself and will at the end connect
circles that intersect each other. This is a major improvement from the
previous version.