Added a new trial.
Trial 7:

only uses 2 shapes and always uses shape1. This trial was done in order to see if less shape sets would improve the accuracy results. This confirms that another machine learning algorithm with svm must be implemented in order to discern which features would be usable as the best descriptor of the different shapes.